rename sim_executor to simulator
