+++ qiskit_ibm_runtime/executor/executor.py
@@ -22,7 +22,7 @@ from ..fake_provider.local_service import QiskitRuntimeLocalService
 from ..options_models.converters import to_runtime_options
 from ..options_models.executor import ExecutorOptions
 from ..quantum_program.params_converters import QUANTUM_PROGRAM_PARAMS_CONVERTERS
-from ..sim_executor import SimRuntimeJob
+from ..simulator import SimRuntimeJob
 from ..utils.default_session import get_cm_session
 
 if TYPE_CHECKING:
