+++ test/unit/simulator/test_run_quantum_program.py
@@ -32,7 +32,7 @@ from samplomatic.transpiler import generate_boxing_pass_manager
 
 from qiskit_ibm_runtime.fake_provider.backends.fez import FakeFez
 from qiskit_ibm_runtime.quantum_program import QuantumProgram
-from qiskit_ibm_runtime.sim_executor.run_quantum_program import run_quantum_program
+from qiskit_ibm_runtime.simulator.run_quantum_program import run_quantum_program
 
 from ...ibm_test_case import IBMTestCase
